Add active degrees-of-freedom mask to Grid
To avoid communication of the auxiliary state, we need to update the
auxiliary state for ghost degrees-of-freedom.  The problem is some ghost
degrees-of-freedom contain valid data and some don't and the physics
routines don't work with invalid data.  So, we need to be able to mask
out invalid degrees-of-freedom.

The test of the mask array is added to an existing test to avoid the
cost of launching another MPI job in the test suite.